SSO_Backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit==2.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.4.0
django-ratelimit==4.1.0
drf-yasg==1.21.7